feat: Minimal project skeleton
The project is equipped with setuptools and nox, so it is possible
to test. We use black, pylint, flake8, isort, behave for testing.

The project uses github workflows to automatically run nox
for testing.

The application works with the configuration, if the configuration
is missing, it writes the default one. By default, the existence
of directories is verified, and if they are missing, they are also
created.

Standard directories are: ./characters, ./modules.