feat!:[PoC] System for generic rule modules
problem: Currently, we had a modules with python code for rule modules.
But we can write more generic app for working with generic modules
describing rules written in .toml files.

This is working Proof of Concept and it is a base for future
development. The work is based on [1].

What should be next?
* to finish a PoC to be ready to print first part of character sheet
* write down documentation for modules

[1] https://github.com/celestian/rpgt/blob/main/docs/design/generic_character_creation.md

Resolves #2 #5