feat!: Load character
This is another big commit. It seems that rpgt has proper inner logic
and strcuture. It also behaves as I am expecting.

The issue is, there is no documentation, no examples, no tests.

So, the next step is to stabiize the code:
* tests
* examples
* documentation

After those steps, commits should not be so crazy, as they are right
now.